requirements.txt
asgiref==3.8.1
Django==5.1.4
django-cors-headers==4.3.1
django-filter==24.3
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
drf-yasg==1.21.8
inflection==0.5.1
packaging==24.2
psycopg==3.2.3
psycopg2==2.9.10
PyJWT==2.10.1
pytz==2024.2
PyYAML==6.0.2
simplejwt==2.0.1
sqlparse==0.5.3
typing==3.7.4.3
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
gunicorn==21.2.0
